feat!: remove login analytics and default skill sync
Deletes the last two consumers of the duet.so web origin. `duet login`
still runs the device flow and writes DUET_API_KEY, but no longer posts
the cli_login analytics event or syncs default skills; the on-load
auto-sync in run/rpc is gone too. --skip-skill-sync/--no-skill-sync are
tolerated as deprecated no-ops so existing scripts do not break.
duet-app-url.ts collapses to duet-api-url.ts (DUET_APP_BASE_URL is
removed; DUET_API_BASE_URL stays for device login and feedback).

Version 0.2.0 for the breaking API-surface change.
