+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@duetso/agent",
-  "version": "0.1.202",
+  "version": "0.2.0",
   "description": "An opinionated full-stack agent turn runner with native memories, interrupts, and multi-agent orchestration",
   "keywords": [
     "agent",
